+++ src/xulbux/console.py
@@ -9,7 +9,7 @@ from .base.consts import COLOR, CHARS, ANSI
 
 from .format_codes import _PATTERNS as _FC_PATTERNS, FormatCodes  # type: ignore[private-access]
 from .string import String
-from .color import Color, hexa
+from .color import Color
 from .regex import LazyRegex
 
 from typing import ValuesView, Generator, Callable, KeysView, Optional, Literal, TypeVar, TextIO, Any, overload, cast
@@ -380,30 +380,27 @@ class Console(metaclass=_ConsoleMeta):
         format_linebreaks: bool = True,
         start: str = "",
         end: str = "\n",
-        title_bg_color: Optional[Rgba | Hexa] = None,
+        title_bg_color: Optional[str | Rgba | Hexa] = None,
         default_color: Optional[Rgba | Hexa] = None,
         tab_size: int = 8,
         title_px: int = 1,
         title_mx: int = 2,
     ) -> None:
         """Prints a nicely formatted log message.\n
-        -------------------------------------------------------------------------------------------
+        ---------------------------------------------------------------------------------------------
         - `title` -⠀the title of the log message (e.g. `DEBUG`, `WARN`, `FAIL`, etc.)
         - `prompt` -⠀the log message
         - `format_linebreaks` -⠀whether to format (indent after) the line breaks or not
         - `start` -⠀something to print before the log is printed
         - `end` -⠀something to print after the log is printed (e.g. `\\n`)
-        - `title_bg_color` -⠀the background color of the `title`
-        - `default_color` -⠀the default text color of the `prompt`
+        - `title_bg_color` -⠀the background color of the `title` (console color name, RGBA, or HEXA)
+        - `default_color` -⠀the default text color of the `prompt` (RGBA or HEXA)
         - `tab_size` -⠀the tab size used for the log (default is 8 like console tabs)
         - `title_px` -⠀the horizontal padding (in chars) to the title (if `title_bg_color` is set)
         - `title_mx` -⠀the horizontal margin (in chars) to the title\n
-        -------------------------------------------------------------------------------------------
+        ---------------------------------------------------------------------------------------------
         The log message can be formatted with special formatting codes. For more detailed
         information about formatting codes, see `format_codes` module documentation."""
-        has_title_bg: bool = False
-        if title_bg_color is not None and (Color.is_valid_rgba(title_bg_color) or Color.is_valid_hexa(title_bg_color)):
-            title_bg_color, has_title_bg = Color.to_hexa(title_bg_color), True
         if tab_size < 0:
             raise ValueError("The 'tab_size' parameter must be a non-negative integer.")
         if title_px < 0:
@@ -411,8 +408,18 @@ class Console(metaclass=_ConsoleMeta):
         if title_mx < 0:
             raise ValueError("The 'title_mx' parameter must be a non-negative integer.")
 
+        title_fg: str = "_c"
+        has_title_bg: bool = False
         title = "" if title is None else title.strip().upper()
-        title_fg = Color.text_color_for_on_bg(cast(hexa, title_bg_color)) if has_title_bg else "_color"
+
+        if title_bg_color is not None:
+            if str(title_bg_color).lower() in ANSI.POSSIBLE_COLOR_MAP:
+                has_title_bg = True
+                title_fg = "black"
+            elif Color.is_valid_rgba(title_bg_color) or Color.is_valid_hexa(title_bg_color):
+                has_title_bg = True
+                title_bg_color = Color.to_hexa(title_bg_color)
+                title_fg = str(Color.text_color_for_on_bg(title_bg_color))
 
         px, mx = (" " * title_px) if has_title_bg else "", " " * title_mx
         tab = " " * (tab_size - 1 - ((len(mx) + (title_len := len(title) + 2 * len(px))) % tab_size))
@@ -435,8 +442,8 @@ class Console(metaclass=_ConsoleMeta):
             )
         else:
             FormatCodes.print(
-                f"{start}{mx}[bold][{title_fg}]{f'[BG:{title_bg_color}]' if title_bg_color else ''}{px}{title}{px}[_]{mx}"
-                + f"{tab}{f'[{default_color}]' if default_color else ''}{prompt}[_]",
+                f"{start}{mx}[b|{title_fg}{f'|bg:{title_bg_color}' if has_title_bg else ''}]{px}{title}{px}[_]{mx}"
+                f"{tab}{f'[{default_color}]' if default_color else ''}{prompt}[_]",
                 default_color=default_color,
                 end=end,
             )
@@ -467,7 +474,7 @@ class Console(metaclass=_ConsoleMeta):
                 format_linebreaks=format_linebreaks,
                 start=start,
                 end=end,
-                title_bg_color=COLOR.YELLOW,
+                title_bg_color="br:yellow",
                 default_color=default_color,
             )
             cls.pause_exit("", pause=pause, exit=exit, exit_code=exit_code, reset_ansi=reset_ansi)
@@ -495,7 +502,7 @@ class Console(metaclass=_ConsoleMeta):
             format_linebreaks=format_linebreaks,
             start=start,
             end=end,
-            title_bg_color=COLOR.BLUE,
+            title_bg_color="br:blue",
             default_color=default_color,
         )
         cls.pause_exit("", pause=pause, exit=exit, exit_code=exit_code, reset_ansi=reset_ansi)
@@ -523,7 +530,7 @@ class Console(metaclass=_ConsoleMeta):
             format_linebreaks=format_linebreaks,
             start=start,
             end=end,
-            title_bg_color=COLOR.TEAL,
+            title_bg_color="br:green",
             default_color=default_color,
         )
         cls.pause_exit("", pause=pause, exit=exit, exit_code=exit_code, reset_ansi=reset_ansi)
@@ -579,7 +586,7 @@ class Console(metaclass=_ConsoleMeta):
             format_linebreaks=format_linebreaks,
             start=start,
             end=end,
-            title_bg_color=COLOR.RED,
+            title_bg_color="br:red",
             default_color=default_color,
         )
         cls.pause_exit("", pause=pause, exit=exit, exit_code=exit_code, reset_ansi=reset_ansi)
@@ -607,7 +614,7 @@ class Console(metaclass=_ConsoleMeta):
             format_linebreaks=format_linebreaks,
             start=start,
             end=end,
-            title_bg_color=COLOR.MAGENTA,
+            title_bg_color="br:magenta",
             default_color=default_color,
         )
         cls.pause_exit("", pause=pause, exit=exit, exit_code=exit_code, reset_ansi=reset_ansi)
@@ -625,22 +632,29 @@ class Console(metaclass=_ConsoleMeta):
         indent: int = 0,
     ) -> None:
         """Will print a box with a colored background, containing a formatted log message.\n
-        -------------------------------------------------------------------------------------
+        ---------------------------------------------------------------------------------------
         - `*values` -⠀the box content (each value is on a new line)
         - `start` -⠀something to print before the log box is printed (e.g. `\\n`)
         - `end` -⠀something to print after the log box is printed (e.g. `\\n`)
-        - `box_bg_color` -⠀the background color of the box
+        - `box_bg_color` -⠀the background color of the box (console color name, RGBA, or HEXA)
         - `default_color` -⠀the default text color of the `*values`
         - `w_padding` -⠀the horizontal padding (in chars) to the box content
         - `w_full` -⠀whether to make the box be the full console width or not
         - `indent` -⠀the indentation of the box (in chars)\n
-        -------------------------------------------------------------------------------------
+        ---------------------------------------------------------------------------------------
         The box content can be formatted with special formatting codes. For more detailed
         information about formatting codes, see `format_codes` module documentation."""
         if w_padding < 0:
             raise ValueError("The 'w_padding' parameter must be a non-negative integer.")
         if indent < 0:
             raise ValueError("The 'indent' parameter must be a non-negative integer.")
+
+        if str(box_bg_color).lower() in ANSI.POSSIBLE_COLOR_MAP:
+            pass
+        elif Color.is_valid_rgba(box_bg_color) or Color.is_valid_hexa(box_bg_color):
+            box_bg_color = Color.to_hexa(box_bg_color)
+        else:
+            raise ValueError("The 'box_bg_color' parameter must be a valid console color name, RGBA value, or HEXA value.")
 
         if Color.is_valid(box_bg_color):
             box_bg_color = Color.to_hexa(box_bg_color)
@@ -1720,7 +1734,7 @@ class ProgressBar:
 
         self._current_progress_str = progress_text
         self._last_line_len = len(progress_text)
-        self._original_stdout.write(f"\r{progress_text}")
+        self._original_stdout.write(f"{ANSI.CHAR}[2K\r{progress_text}")
         self._original_stdout.flush()
 
     def _get_formatted_info_and_bar_width(